refactor(product-components): fix circular type imports
